renamed Client to LegacyClient, Client is an alias to LegacyClient for backward compatibility
